requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis and Task Queue
redis==5.0.1
celery==5.3.6
flower==2.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0
pyjwt==2.8.0

# API Clients
google-generativeai==0.3.2
google-cloud-aiplatform==1.42.1
openai==1.12.0
anthropic==0.18.1
httpx==0.26.0
aiohttp==3.9.3

# AWS Services
boto3==1.34.34
botocore==1.34.34

# Pydantic and Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.1
pillow==10.2.0
python-magic==0.4.27

# Rate Limiting
slowapi==0.1.9

# Logging and Monitoring
structlog==24.1.0

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Type hints
types-redis==4.6.0.20240106
types-boto3==1.0.2
